pyproject.toml
[project]
authors = [
  {name = "seonglae", email = "sungle3737@gmail.com"},
]
dependencies = [
    "fire>=0.5.0",
    "transformers>=4..0",
    "torch==2.1.0+cu118",
    "faiss-cpu>=1.7.4",
    "datasets>=2.14.5",
    "huggingface_hub>=0.17.0",
    "pymilvus>=2.3.2",
    "tiktoken>=0.5.1",
    "teicli>=0.3.3",
    "python-dotenv>=1.0.0",
    "evaluate[evaluator]>=0.4.1",
    "pyarrow>=14.0.1",
    "openai>=1.3.5",
    "urllib3<=2",
    "types-requests>=2.31.0.10",
    "accelerate>=0.24.1",
]
description = "Add your description here"
name = "resrer"
readme = "README.md"
requires-python = ">= 3.8"
version = "0.1.0"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.rye]
dev-dependencies = [
    "autopep8>=2.0.4",
    "pip>=23.2.1",
    "mypy>=1.6.0",
    "setuptools>=68.2.2",
]
managed = true

[tool.hatch.metadata]
allow-direct-references = true

[[tool.rye.sources]]
name = "cuda"
type = "index"
url = "https://download.pytorch.org/whl/cu118"
